+++ docs/.vitepress/theme/index.ts
@@ -1,0 +1,12 @@
+import DefaultTheme from 'vitepress/theme'
+import './custom.css'
+import DemoPreview from './components/DemoPreview.vue'
+import CdnNotice from './components/CdnNotice.vue'
+
+export default {
+	...DefaultTheme,
+	enhanceApp({ app }) {
+		app.component('DemoPreview', DemoPreview)
+		app.component('CdnNotice', CdnNotice)
+	},
+}